Improve exceptions handling and move models in a package
