+++ pdfact-api/src/main/java/pdfact/api/model/RequestPayload.java
@@ -1,0 +1,33 @@
+package pdfact.api.model;
+
+import java.util.List;
+
+public class RequestPayload {
+    private String url;
+    private String unit;
+    private List<String> roles;
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+
+    public String getUnit() {
+        return unit;
+    }
+
+    public void setUnit(String unit) {
+        this.unit = unit;
+    }
+
+    public List<String> getRoles() {
+        return roles;
+    }
+
+    public void setRoles(List<String> roles) {
+        this.roles = roles;
+    }
+}
